+++ src/components/BankTransactionsMobileList/BankTransactionsMobileListSplitAndMatchForm.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 import { type BankTransaction } from '@internal-types/bankTransactions'
 import { hasMatch } from '@utils/bankTransactions/shared'
 import { BankTransactionsMobileListMatchForm } from '@components/BankTransactionsMobileList/BankTransactionsMobileListMatchForm'
-import { BankTransactionsMobileListSplitForm } from '@components/BankTransactionsMobileList/BankTransactionsMobileListSplitForm'
+import { BankTransactionsMobileListSplitForm } from '@components/BankTransactionsMobileList/BankTransactionsMobileListSplitForm/BankTransactionsMobileListSplitForm'
 import { TextButton } from '@components/Button/TextButton'
 
 interface BankTransactionsMobileListSplitAndMatchFormProps {
